refactor(racing-car): RacingService -> RacingGame 으로 변경, 경주 결과를 출력하는 역할을 도메인이 수행하지 않도록 수정
