+++ src/main/java/racingcar/RacingCarMain.java
@@ -3,7 +3,7 @@ package racingcar;
 import java.util.List;
 import java.util.Scanner;
 import racingcar.domain.Car;
-import racingcar.service.RacingService;
+import racingcar.domain.RacingGame;
 import racingcar.ui.RacingCarInput;
 import racingcar.ui.RacingCarOutput;
 
@@ -13,9 +13,16 @@ public class RacingCarMain {
     public static void main(String[] args) {
         List<Car> cars = RacingCarInput.carNameInput();
         int round = RacingCarInput.roundInput();
-        RacingService racingService = new RacingService(cars);
+        RacingGame racingGame = new RacingGame(cars);
         RacingCarOutput.printRoundResultMessage();
-        racingService.race(round);
-        RacingCarOutput.printWinners(racingService.findWinners());
+        raceAndPrintResult(round, racingGame);
+        RacingCarOutput.printWinners(racingGame.findWinners());
+    }
+
+    private static void raceAndPrintResult(int round, RacingGame racingGame) {
+        for (int i = 0; i < round; i++) {
+            List<Car> raceResult = racingGame.race();
+            RacingCarOutput.printRoundResult(raceResult);
+        }
     }
 }
